refact: move BaseEntity to entity package
