+++ post/src/main/kotlin/com/bibbi/infrastructure/jpa/entity/MemberRealEmojiEntity.kt
@@ -1,6 +1,6 @@
 package com.bibbi.infrastructure.jpa.entity
 
-import com.bibbi.domain.BaseEntity
+import com.bibbi.domain.entity.BaseEntity
 import com.bibbi.domain.entity.Emoji
 import com.bibbi.domain.entity.MemberRealEmoji
 import jakarta.persistence.*
